added entities for orders

+++ src/main/java/com/ngon/backend/dto/ProductRequest.java
@@ -2,10 +2,12 @@ package com.ngon.backend.dto;
 
 import jakarta.validation.constraints.NotBlank;
 
+import java.math.BigDecimal;
+
 public record ProductRequest(
         @NotBlank String name,
         @NotBlank int quantity,
-        @NotBlank float price,
+        @NotBlank BigDecimal price,
         @NotBlank String category
 )
 {}
