+++ src/main/java/com/ngon/backend/entity/Product.java
@@ -2,6 +2,8 @@ package com.ngon.backend.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "products")
 public class Product
@@ -14,17 +16,17 @@ public class Product
     @Column(nullable = false)
     private int quantity;
     @Column(nullable = false)
-    private float price;
+    private BigDecimal price;
     @Column(nullable = false)
     private String category;
 
     public void setName(String name) {this.name = name;}
     public void setQuantity(int quantity) {this.quantity = quantity;}
-    public void setPrice(float price) {this.price = price;}
+    public void setPrice(BigDecimal price) {this.price = price;}
     public void setCategory(String category) {this.category = category;}
 
     public String getName() {return this.name;}
     public int getQuantity() {return this.quantity;}
-    public float getPrice() {return this.price;}
+    public BigDecimal getPrice() {return this.price;}
     public String getCategory() {return this.category;}
 }
